Fix bug in widening-multiply-and-accumulate optimization.
GCC Bugzilla 50717.

Backport from FSF.

+++ gcc/testsuite/gcc.target/arm/wmul-8.c
@@ -4,7 +4,7 @@
 long long
 foo (long long a, int *b, int *c)
 {
-  return a + *b * *c;
+  return a + (long long)*b * *c;
 }
 
 /* { dg-final { scan-assembler "smlal" } } */
